# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: OmniRepo API
  version: "1.0.0"
  description: |
    REST API for OmniRepo — self-hosted artifact repository server.
    Serves OCI/Docker, RPM, APT/Debian, PyPI, Helm, RAW, S3, and Git
    from a single container on one HTTP/HTTPS port.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ApiErrorClass:
      type: string
      enum: [validation, permission, transient, operator_action_required]
      description: Stable client-facing error classification — drives UI icon/copy/CTA.
    ApiErrorEnvelope:
      type: object
      required: [code, message, class]
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*\\.[a-z][a-z0-9_]*$"
          maxLength: 80
          description: Stable machine identifier (e.g. "repo.not_found")
        message:
          type: string
          maxLength: 280
          description: Human-friendly single-sentence message (no internal paths)
        hint:
          type: string
          maxLength: 280
          description: Optional single-sentence remediation guidance
        class:
          $ref: '#/components/schemas/ApiErrorClass'
        incident_id:
          type: string
          maxLength: 64
          description: Server-generated correlation ID (UUID v7) — matches slog request_id and the X-Incident-Id response header
        details:
          type: object
          additionalProperties: true
          properties:
            field:
              type: string
              description: Dot-path of the offending form field (validation class)
            fields:
              type: object
              additionalProperties:
                type: string
              description: Map of field-path to error-code for multi-field validation failures
            retry_after_ms:
              type: integer
              minimum: 0
              description: Suggested retry delay in ms (transient class)
            operator_route:
              type: string
              description: UI path (e.g. "/admin/trivy") for operator_action_required deep-link
            operator_label:
              type: string
              description: CTA label for the operator deep-link (e.g. "Go to Admin → Trivy")
...
    Repo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
        type:
          type: string
        name:
          type: string
        description_md:
          type: string
        auto_scan:
          type: boolean
        block_on_severity:
          type: string
          enum: [none, low, medium, high, critical]
        public_read:
          type: boolean
        size_bytes:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    RepoCreate:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        description_md:
          type: string
        auto_scan:
          type: boolean
        block_on_severity:
          type: string
          enum: [none, low, medium, high, critical]
        public_read:
          type: boolean
    RepoPatch:
      type: object
      properties:
        description_md:
          type: string
        auto_scan:
          type: boolean
        block_on_severity:
          type: string
          enum: [none, low, medium, high, critical]
        public_read:
          type: boolean
...
    StatusOK:
      type: object
      properties:
        status:
          type: string

    # -- Scans --
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    cursorParam:
      name: cursor
      in: query
      schema:
        type: string
paths:
...
  /projects/{name}/repos:
    get:
      operationId: listRepos
      tags: [repos]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/cursorParam'
      responses:
        "200":
          description: Repos in project
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Repo'
                  next_cursor:
                    type: string
                    nullable: true
        default:
          $ref: '#/components/responses/ValidationError'
    post:
      operationId: createRepo
      tags: [repos]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoCreate'
      responses:
        "200":
          description: Repo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repo'
        default:
          $ref: '#/components/responses/ValidationError'
  /projects/{name}/repos/{type}/{repo}:
    get:
      operationId: getRepo
      tags: [repos]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repo detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repo'
        default:
          $ref: '#/components/responses/ValidationError'
    patch:
      operationId: patchRepo
      tags: [repos]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoPatch'
      responses:
        "200":
          description: Repo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repo'
        default:
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: deleteRepo
      tags: [repos]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repo soft-deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOK'
        default:
          $ref: '#/components/responses/ValidationError'
...